requirements.txt
# requirements.txt for BDH-Official project

# IMPORTANT: Install PyTorch separately based on your GPU:
#
# For GTX 1060 and older GPUs (compute capability 6.1+):
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#
# For newer GPUs (RTX 30xx/40xx/50xx, compute capability 7.0+):
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
#
# For CPU-only installation:
#   pip install torch torchvision torchaudio
#
# PyTorch is not included here to allow flexible CUDA version selection
# After installing PyTorch, install remaining dependencies with:
#   pip install -r requirements.txt

# Data and Metrics
numpy>=1.21.0,<2.0
pandas>=1.3.0
scikit-learn>=1.0.0
datasets>=2.0.0
tiktoken>=0.5.0

# Visualization and Utilities
matplotlib>=3.5.0
seaborn>=0.12.0
tqdm>=4.64.0
requests>=2.28.0

# Data Ingestion
pyarrow>=10.0.0  # Parquet file support
pyyaml>=6.0
python-dateutil>=2.8.0
